Require py27-dpkt as part of the test image
In short, tests/sys/opencrypto/runtests has a copy of the dpkt.py module
checked in from the dpkt package, which is based on a 2014 snapshot of
the package. In order to make forward progress and eliminate the
duplicate copy of dpkt, the first logical step is to require the upstream
package instead of relying on a partial copy of the dpkt package.

PR:	237403

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>